# api/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeERP API
  version: 0.1.0
  description: >
    ForgeERP modernized ERP/CRM API (Dolibarr functional successor).
    This document is authoritative for REST surface; phase routers must register
    every path declared here.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/finance/entries:
    post:
      operationId: postEntry
      summary: Validate, chain-link and post an entry (requires finance.entry.write)
      responses:
        '201': { description: Posted }
        '422': { description: Unbalanced or locked year }
...